requirements.txt
# ===========================================
# 投资理财管理系统 - Python 依赖
# ===========================================
# 安装命令: pip install -r requirements.txt
# Python 版本要求: 3.8+

# -------------------------------------------
# Web 框架
# -------------------------------------------
Flask>=2.0.0
Flask-CORS>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0

# -------------------------------------------
# 数据库
# -------------------------------------------
SQLAlchemy>=1.4.0
PyMySQL>=1.0.0
cryptography>=3.0.0

# -------------------------------------------
# 配置管理
# -------------------------------------------
PyYAML>=6.0

# -------------------------------------------
# 请求验证
# -------------------------------------------
pydantic>=2.0.0

# -------------------------------------------
# API 限流
# -------------------------------------------
Flask-Limiter>=3.0.0

# -------------------------------------------
# 认证安全
# -------------------------------------------
PyJWT>=2.8.0
bcrypt>=4.0.0

# -------------------------------------------
# 生产服务器
# -------------------------------------------
gunicorn>=21.0.0

# -------------------------------------------
# 大模型 API（AI分析功能）
# -------------------------------------------
openai>=1.0.0
anthropic>=0.18.0

# -------------------------------------------
# 金融数据源
# -------------------------------------------
akshare>=1.12.0
baostock>=0.8.8
tushare>=1.3.0
requests>=2.28.0

# -------------------------------------------
# 数据处理
# -------------------------------------------
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# -------------------------------------------
# 可选：开发测试
# -------------------------------------------
# pytest>=7.0.0
# black>=23.0.0